Render Component Lifecycles
Learning where you have to use specific reactive features differently in React, compared to SolidJS, since their lifecycles are a bit different in a few aspects.

This does allow the demo to work!

I still have to figure out how to append to an existing component, for the temporary FormatOptions form element handling.

https://stackoverflow.com/questions/38256332/in-react-whats-the-difference-between-onchange-and-oninput
https://react.dev/reference/react-dom/client/createRoot
https://react.dev/learn/rendering-lists#keeping-list-items-in-order-with-key
https://stackoverflow.com/questions/38192552/react-jsx-append-html-react-object-after-initialization