NBTify Upgrade - 1.90.0
Working on bringing the latest changes over from NBTify to Dovetail!

Things aren't completely compatible yet though, so it's going to take some work. The main challenge has to do with the re-typing of the BedrockLevel format option, which is now a boolean rather than a specific value. It also involves more in-depth validation against the incoming root NBT object, since it needs to have a `StorageVersion` key with a number value, which is what should be used for the file header itself.