NBTify Revert
Going to rework some of the newer changes to NBTify, it isn't completely working nicely with Dovetail. This commit brings Dovetail back to the previous NBTify version before the BedrockLevel precision update (where the BedrockLevel use to work with any file, even when it technically wouldn't make any sense to occur there).

Should the BedrockLevel header still be applicable to any file, even if it doesn't use the same format as vanilla Bedrock? Say for example, compressed big endian, should NBTify support reading and writing that with a BedrockLevel header? Support-wise, it's easy to implement, but I don't want it to accidentally allow people to make files they didn't mean to, because they might only set the BedrockLevel header for example, but they need to update the endian type and remove the compression, for example, for it to work with Bedrock itself.