Vector NBT Icons!
Making these with Figma! They aren't quite done yet, but here's the initial versions. I want to save these as vectors as well.

Just realized a neat idea, maybe the icons for NBT types can use the same HEX colors of the different wool types in game! Not that it makes sense particularly, but it could be a cool crossover
This image doesn't do that, thought I'd share what my demos look like so far though

I still have to adjust the formatting and the colors but here's what I have so far

I want to make proper glyphs for the fancy tags as well.