Strict Typings + Key Escapes
Strict typings are what I've been doing lately too, it's not for the purpose of always doing it just because, and to have everything 'complete', but more for the validation of intent for how the code is meant to work, almost like breakpoints throughout the codebase if something somewhere is expected to work in a certain way.

Also copied over the edits to the `escapeString()` function, this is for tree view key name rendering.