Main Show Renderer
Going to start refactoring things a bit to make the code more easy to read, even if it might make it less React-ish out of the box. That was the outlook I had originally, to where I wanted to keep the syntax the same for working directly with React or Solid out of the box, without needing to edit much explicitly. Going to move on from that, and instead go all in on Solid for this project.